api/requirements.txt
# Dr Green API Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0

# Utilities
pydantic==2.5.0
